pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sincpd"
version = "0.1.0"
description = "Explainable Sinc-based PD diagnosis & severity from vGRF (PyTorch)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Mahan Veisi", email = "example@example.com" }]
license = { file = "LICENSE" }
dependencies = [
  "torch>=2.2",
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "scikit-learn>=1.3",
  "tqdm>=4.66",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pre-commit>=3.8",
  "black>=24.8",
  "isort>=5.13",
  "flake8>=7.1",
  "pytest-cov>=5.0"
]

[project.scripts]
sincpd = "sincpd.cli:main"

[tool.isort]
profile = "black"
